Add Profile components and several page changes
