+++ client/src/scss/components/_header.scss
@@ -62,11 +62,13 @@ nav > div {
 }
   
 /* Links inside the dropdown */
-.dropdown-content a {
-    color: $nh-white;
-    padding: 12px 16px;
-    text-decoration: none;
-    display: block;
+.dropdown-content {
+    > button, a {
+        color: $nh-white;
+        padding: 12px 16px;
+        text-decoration: none;
+        display: block;
+    }
 } 
   
 /* Show the dropdown menu on hover */
