+++ client/src/scss/utilities/_mixins.scss
@@ -9,17 +9,9 @@
     margin-left: auto;
 }
 
-@mixin button-properties {
+@mixin button-properties($bg-color) {
     border: none;
-    background-color: $nh-button;
-    border-radius: 16px;
-    height: 50px;
-    cursor: pointer;
-}
-
-@mixin alt-button-properties {
-    border: none;
-    background-color: white;
+    background-color: #{$bg-color};
     border-radius: 16px;
     height: 50px;
     cursor: pointer;
